searchbar corrections
