+++ app/javascript/plugins/init_mapbox.js
@@ -191,7 +191,6 @@ const initMapbox = () => {
         const el = document.createElement('div');
         const category = marker.category;
         const color = setColor(marker.average)
-        console.log(marker)
         el.className = `marker-${category}-${color}`;
 
       new mapboxgl.Marker( el )
